bugfix in export of layered textures
git-svn-id: https://svn.intern.artcom.de/pro60/trunk/src@20235 f3027e6b-e3f5-0310-a8e1-edde99107618

+++ Y60/products/maya-xport/MaterialExporter.cpp
@@ -526,7 +526,7 @@ MaterialExporter::exportLayeredTexture(const MFnMesh * theMesh,
 
     // Iterate over all texture layers
     // Maya shades the first texture last, therefore we reverse the order
-    for (unsigned int i = myInputPlug.numElements() - 1; i >= 0; --i) {
+    for (int i = static_cast<int>(myInputPlug.numElements()) - 1; i >= 0; --i) {
         // Export alpha and blend mode for the texture
         float myAlpha = 1.0;
         MPlug myAlphaPlug;
